Add set data structure definition
Previously I was abusing stacks as sets, but this was hard to read
because it was unclear whether the stacks were actually used for their
stack properties or just as a makeshift set.  In addition to that stacks
do not have the idempotence of sets, which as not an issue in practice,
but it could have caused bugs.